Allow min/max filtering on transactions
- Make "credit" a required arg
- Enable decimal presentation of transaction amount while retaining
integer storage

+++ elixir/lib/homework_web/schemas/transactions_schema.ex
@@ -9,7 +9,7 @@ defmodule HomeworkWeb.Schemas.TransactionsSchema do
   object :transaction do
     field(:id, non_null(:id))
     field(:user_id, :id)
-    field(:amount, :integer)
+    field(:amount, :float)
     field(:credit, :boolean)
     field(:debit, :boolean)
     field(:description, :string)
